Update theMoth.py
Changed to support the standard Invisible Sun rolling style. When just told to roll, it will return a random roll from 0-9. When told to roll with a number, will roll a "mundane die" with results 0-9, then roll "magic die" until the total number of dice is reached. This makes it easier to see and track flux, and prevents the momentary hype of seeing a 10 before realizing it's actually a 0. Regular dice rolling functionality is maintained using the unchanged syntax.