Environment Variables and Docker Support
I want to run our bot within a Kubernetes cluster, which means I need
the ability to run a containerized version of the bot. On top of that,
instead of hard-coding the bot token, I've modified the script to
allow specifying it using an environment variable.

As I figured out how to run things, I documented what I figured out in
the README under a new section. This will hopefully make it a bit
easier for the next person who wants to run their own instance.