Provide Game Specific Information Legally
I wanted a way to be able to look up information about various cards
without infringing on MCG's copyright. I can't reproduce the actual
game information, but fortunately they provided a searchable text file
which is parseable with a little effort.

This change allows the bot to work exactly the same if a searchable
text file is not provided, but will parse that file and add additional
commands if it is. The basic proofs of concept are the commands
`/lookup` and `/random`, but I intend them to be stepping stones to
more interesting options.